Don't cleanup files outside the output directory
git-svn-id: https://pyjamas.svn.sourceforge.net/svnroot/pyjamas/trunk@1669 7a2bd370-bda8-463c-979e-2900ccfb811e

+++ pyjs/src/pyjs/browser.py
@@ -88,7 +88,8 @@ class BrowserLinker(linker.BaseLinker):
         self._create_nocache_html()
         if not self.keep_lib_files:
             for fname in self.remove_files:
-                os.unlink(fname)
+                if fname.find(self.output) == 0:
+                    os.unlink(fname)
 
     def merge_resources(self, dir_name):
         if not dir_name in self.merged_public:
